requirements.txt
asgiref==3.3.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.7
django-autoslug==1.9.8
django-crispy-forms==1.11.1
django-import-export==2.5.0
django-ipware==3.0.2
django-js-asset==1.2.2
django-mptt==0.13.2
django-multiselectfield==0.1.12
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
et-xmlfile==1.0.1
gunicorn==20.1.0
install==1.3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.1.2

PyMySQL==1.0.2
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
